Allow pipeline in private or public mode (#36)
- rename text, variables, filenames, etc. to align with
https://github.com/nih-cfde/icc-eval-coordination/pull/2#event-16933083340
- upgrade all package versions. upgrade to modern eslint config syntax.
update tsconfigs.
- update `gather.ts` to allow a split between private/public data

The gist here is this:

We're going to make this repo public again, so anyone can see the code
and the gathered public data. Then a private repo, with barely anything
in it, will checkout this exact code, run it with the `PRIVATE=true` env
variable set from the workflow. When running in "private mode", it will
attempt to load existing data in the public repo (which is already
checked out locally) instead of re-doing actual network queries to get
it. It then also won't re-save the public data, to avoid data
duplication. But it will save the private data gathered into the private
repo. Then, when running in "public mode", the script will work as it
normally did before, except it will not try to get any of the private
stuff.

I've also checked the changelogs of every package I did a major version
upgrade on, and none of the breaking changes affect us (e.g. drop
support for node 18, drop support for Vue 2).

@the-mayer I'd say you only need to check the `gather.ts` file to see if
you think the logic is correct. Don't need to check the package
upgrades. If you want, you can try running it locally in public mode,
but given the time sensitivity, I think it's not necessary, and I can
take responsibility for any errors in the PR.